+++ settingrepo-impl/src/main/java/com/dwarfeng/settingrepo/impl/handler/FormatLocalCacheHandlerImpl.java
@@ -5,91 +5,78 @@ import com.dwarfeng.settingrepo.stack.handler.FormatLocalCacheHandler;
 import com.dwarfeng.settingrepo.stack.handler.Formatter;
 import com.dwarfeng.settingrepo.stack.handler.FormatterHandler;
 import com.dwarfeng.settingrepo.stack.service.SettingCategoryMaintainService;
+import com.dwarfeng.subgrade.impl.handler.Fetcher;
+import com.dwarfeng.subgrade.impl.handler.GeneralLocalCacheHandler;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
 import com.dwarfeng.subgrade.stack.exception.HandlerException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.*;
-import java.util.concurrent.locks.ReadWriteLock;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 @Component
 public class FormatLocalCacheHandlerImpl implements FormatLocalCacheHandler {
 
-    @Autowired
-    private FormatterFetcher formatterFetcher;
+    private final GeneralLocalCacheHandler<StringIdKey, Formatter> handler;
 
-    private final ReadWriteLock lock = new ReentrantReadWriteLock();
-    private final Map<StringIdKey, Formatter> formatterMap = new HashMap<>();
-    private final Set<StringIdKey> notExistSettingCategories = new HashSet<>();
-
-    @Override
-    public Formatter getFormatter(StringIdKey settingCategoryKey) throws HandlerException {
-        try {
-            lock.readLock().lock();
-            try {
-                if (formatterMap.containsKey(settingCategoryKey)) {
-                    return formatterMap.get(settingCategoryKey);
-                }
-                if (notExistSettingCategories.contains(settingCategoryKey)) {
-                    return null;
-                }
-            } finally {
-                lock.readLock().unlock();
-            }
-            lock.writeLock().lock();
-            try {
-                if (formatterMap.containsKey(settingCategoryKey)) {
-                    return formatterMap.get(settingCategoryKey);
-                }
-                if (notExistSettingCategories.contains(settingCategoryKey)) {
-                    return null;
-                }
-                Formatter formatter = formatterFetcher.fetchFormatter(settingCategoryKey);
-                if (Objects.nonNull(formatter)) {
-                    formatterMap.put(settingCategoryKey, formatter);
-                    return formatter;
-                }
-                notExistSettingCategories.add(settingCategoryKey);
-                return null;
-            } finally {
-                lock.writeLock().unlock();
-            }
-        } catch (Exception e) {
-            throw new HandlerException(e);
-        }
+    public FormatLocalCacheHandlerImpl(FormatterFetcher formatterFetcher) {
+        this.handler = new GeneralLocalCacheHandler<>(formatterFetcher);
     }
 
+    @BehaviorAnalyse
     @Override
-    public void clear() {
-        lock.writeLock().lock();
-        try {
-            formatterMap.clear();
-            notExistSettingCategories.clear();
-        } finally {
-            lock.writeLock().unlock();
-        }
+    public boolean exists(StringIdKey key) throws HandlerException {
+        return handler.exists(key);
+    }
+
+    @BehaviorAnalyse
+    @Override
+    public Formatter get(StringIdKey key) throws HandlerException {
+        return handler.get(key);
+    }
+
+    @BehaviorAnalyse
+    @Override
+    public boolean remove(StringIdKey key) {
+        return handler.remove(key);
+    }
+
+    @BehaviorAnalyse
+    @Override
+    public void clear() throws HandlerException {
+        handler.clear();
     }
 
     @Component
-    public static class FormatterFetcher {
+    public static class FormatterFetcher implements Fetcher<StringIdKey, Formatter> {
 
-        @Autowired
-        private SettingCategoryMaintainService settingCategoryMaintainService;
+        private final SettingCategoryMaintainService settingCategoryMaintainService;
 
-        @Autowired
-        private FormatterHandler formatterHandler;
+        private final FormatterHandler formatterHandler;
 
+        public FormatterFetcher(
+                SettingCategoryMaintainService settingCategoryMaintainService,
+                FormatterHandler formatterHandler
+        ) {
+            this.settingCategoryMaintainService = settingCategoryMaintainService;
+            this.formatterHandler = formatterHandler;
+        }
+
+        @Override
         @BehaviorAnalyse
-        @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class)
-        public Formatter fetchFormatter(StringIdKey settingCategoryKey) throws Exception {
-            if (!settingCategoryMaintainService.exists(settingCategoryKey)) {
-                return null;
-            }
-            SettingCategory settingCategory = settingCategoryMaintainService.get(settingCategoryKey);
+        @Transactional(
+                transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class
+        )
+        public boolean exists(StringIdKey key) throws Exception {
+            return settingCategoryMaintainService.exists(key);
+        }
+
+        @Override
+        @BehaviorAnalyse
+        @Transactional(
+                transactionManager = "hibernateTransactionManager", readOnly = true, rollbackFor = Exception.class
+        )
+        public Formatter fetch(StringIdKey key) throws Exception {
+            SettingCategory settingCategory = settingCategoryMaintainService.get(key);
             return formatterHandler.make(settingCategory);
         }
     }