+++ settingrepo-impl/src/main/java/com/dwarfeng/settingrepo/impl/service/telqos/FlcCommand.java
@@ -1,0 +1,92 @@
+package com.dwarfeng.settingrepo.impl.service.telqos;
+
+import com.dwarfeng.settingrepo.stack.handler.Formatter;
+import com.dwarfeng.settingrepo.stack.service.FormatQosService;
+import com.dwarfeng.springtelqos.sdk.command.CliCommand;
+import com.dwarfeng.springtelqos.stack.command.Context;
+import com.dwarfeng.springtelqos.stack.exception.TelqosException;
+import com.dwarfeng.subgrade.stack.bean.key.StringIdKey;
+import org.apache.commons.cli.CommandLine;
+import org.apache.commons.cli.Option;
+import org.apache.commons.lang3.tuple.Pair;
+import org.springframework.stereotype.Component;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+
+@Component
+public class FlcCommand extends CliCommand {
+
+    private static final String COMMAND_OPTION_LOOKUP = "l";
+    private static final String COMMAND_OPTION_CLEAR = "c";
+
+    private static final String[] COMMAND_OPTION_ARRAY = new String[]{
+            COMMAND_OPTION_LOOKUP,
+            COMMAND_OPTION_CLEAR
+    };
+
+    private static final String IDENTITY = "flc";
+    private static final String DESCRIPTION = "格式化器本地缓存操作";
+
+    private static final String CMD_LINE_SYNTAX_LOOKUP = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_LOOKUP) + " id";
+    private static final String CMD_LINE_SYNTAX_CLEAR = IDENTITY + " " +
+            CommandUtil.concatOptionPrefix(COMMAND_OPTION_CLEAR);
+
+    private static final String[] CMD_LINE_ARRAY = new String[]{
+            CMD_LINE_SYNTAX_LOOKUP,
+            CMD_LINE_SYNTAX_CLEAR
+    };
+
+    private static final String CMD_LINE_SYNTAX = CommandUtil.syntax(CMD_LINE_ARRAY);
+
+    private final FormatQosService formatQosService;
+
+    public FlcCommand(FormatQosService formatQosService) {
+        super(IDENTITY, DESCRIPTION, CMD_LINE_SYNTAX);
+        this.formatQosService = formatQosService;
+    }
+
+    @Override
+    protected List<Option> buildOptions() {
+        List<Option> list = new ArrayList<>();
+        list.add(Option.builder(COMMAND_OPTION_LOOKUP).desc("查询格式化器").hasArg().type(String.class).build());
+        list.add(Option.builder(COMMAND_OPTION_CLEAR).desc("清除格式化器").build());
+        return list;
+    }
+
+    @SuppressWarnings("DuplicatedCode")
+    @Override
+    protected void executeWithCmd(Context context, CommandLine cmd) throws TelqosException {
+        try {
+            Pair<String, Integer> pair = CommandUtil.analyseCommand(cmd, COMMAND_OPTION_ARRAY);
+            if (pair.getRight() != 1) {
+                context.sendMessage(CommandUtil.optionMismatchMessage(COMMAND_OPTION_ARRAY));
+                context.sendMessage(CMD_LINE_SYNTAX);
+                return;
+            }
+            switch (pair.getLeft()) {
+                case COMMAND_OPTION_LOOKUP:
+                    handleLookup(context, cmd);
+                    break;
+                case COMMAND_OPTION_CLEAR:
+                    formatQosService.clearLocalCache();
+                    context.sendMessage("本地缓存已清除");
+                    break;
+            }
+        } catch (Exception e) {
+            throw new TelqosException(e);
+        }
+    }
+
+    private void handleLookup(Context context, CommandLine cmd) throws Exception {
+        StringIdKey settingCategoryKey = new StringIdKey((String) cmd.getParsedOptionValue(COMMAND_OPTION_LOOKUP));
+        Formatter formatter = formatQosService.getFormatter(settingCategoryKey);
+        if (Objects.isNull(formatter)) {
+            context.sendMessage("not exists");
+        } else {
+            context.sendMessage(formatter.toString());
+        }
+    }
+}